[DataCollection] fixed an upload bug and changed the description for editing records.
git-svn-id: http://svn.ilias.de/svn/ilias/branches/Release_4_3_x_branch@40589 21b2c9ec-7c21-0410-8b45-9bfb8ed2bfc5

+++ Modules/DataCollection/classes/class.ilDataCollectionRecord.php
@@ -521,8 +521,10 @@ class ilDataCollectionRecord
 	 */
 	public function deleteFile($obj_id)
 	{
-		$file = new ilObjFile($obj_id, false);
-		$file->delete();
+        if(ilObject2::_lookupObjId($obj_id)){
+		    $file = new ilObjFile($obj_id, false);
+		    $file->delete();
+        }
 	}
 
 
@@ -531,8 +533,10 @@ class ilDataCollectionRecord
       */
     public function deleteMob($obj_id)
     {
-        $mob = new ilObjMediaObject($obj_id);
-        $mob->delete();
+        if(ilObject2::_lookupObjId($obj_id)){
+            $mob = new ilObjMediaObject($obj_id);
+            $mob->delete();
+        }
     }
 	
 	/*
